Student_details_app/requirements.txt
tk
numpy==1.24.3
pandasql==0.7.3
Pillow==9.4.0
pip==22.0.4
psycopg2-binary==2.9.3

